fix: resolve all compiler warnings (zero-warning build)
- crackalack_gen.c: cast size_t/unsigned long args to uint64_t for
  PRIu64 format specifiers (gws, kernel sizes, strlen, sizeof)
- crackalack_lookup.c: same casts for gws, sizeof(gpu_ulong) expressions
- rtc_decompress.c: cast sizeof expression to uint64_t
- crackalack_unit_tests.c: move 'err' variable into the #else (OpenCL)
  block where it is actually used; eliminated unused-variable warning
  in Metal builds
- test_index_to_plaintext.c: cast strlen() results to uint64_t
- test_index_to_plaintext_ntlm9.c: same

+++ rtc_decompress.c
@@ -62,7 +62,7 @@ int rtc_decompress(char *filename, uint64_t **ret_uncompressed_table, unsigned i
   /*printf("Total chains in table: %u\n", total_chains_in_table);*/
   uncompressed_table = calloc(num_chains, sizeof(uint64_t) * 2);
   if (uncompressed_table == NULL) {
-    fprintf(stderr, "Error: could not allocate %"PRIu64" bytes in memory for uncompressed table.\n", num_chains * sizeof(uint64_t) * 2);
+    fprintf(stderr, "Error: could not allocate %"PRIu64" bytes in memory for uncompressed table.\n", (uint64_t)(num_chains * sizeof(uint64_t) * 2));
     ret = -2;
     goto done;
   }
